chore: move config loader to core module and improve error handling
